Time: 1 ms (97.68%), Space: 42.5 MB (79.58%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,19 +1,20 @@
 class Solution {
     public int mySqrt(int x) {
         if(x==0) return 0;
-        int lo=1;
-        int hi=x;
-        int ans = 1;
+        
+        long lo=1;
+        long hi=x;
+        long ans = 1;
 
         while(lo<=hi){
-            int mid = lo+(hi-lo)/2;
-            if(((long)mid*mid<=x)){
+            long mid = lo+(hi-lo)/2;
+            if((mid*mid<=x)){
                 ans=mid;
                 lo=mid+1;
             }else{
                 hi=mid-1;
             }
         }
-        return ans ;
+        return (int) ans ;
     }
 }
